refactor: moved ErrorCheckButton and Pagination out of MainPage to separate components
