+++ src/components/Pagination/Pagination.css
@@ -1,0 +1,12 @@
+.pagination {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 2rem;
+}
+
+.currentPage {
+  font-weight: bolder;
+  color: #3f449e;
+  font-size: larger;
+}
