+++ src/components/Pagination/Pagination.types.ts
@@ -1,0 +1,8 @@
+export interface PaginationProps {
+  currentPage: number;
+  totalPages: number;
+  handlePageChange: (
+    newPage: number,
+    event: React.MouseEvent<HTMLButtonElement, MouseEvent>,
+  ) => void;
+}
